Created stage 3 folders and files
