+++ config/algorithm_lists/seaice_stage_3/seaice_stage_3_A001.yml
@@ -1,0 +1,14 @@
+# Algorithm list for chain: seaice
+
+# Optional file selector modules
+l1b_file_selectors:
+#  - your_finder
+#  - your_finder2
+#  - ...
+
+# List of algorithms to call in order
+algorithms:
+  - alg_add_cell_area
+  - alg_add_ice_extent
+  - alg_add_ocean_frac
+  - alg_add_region_mask
