+++ src/clev2er/algorithms/seaice_stage_3/alg_add_cell_area.py
