+++ src/clev2er/algorithms/seaice_stage_3/tests/test_alg_add_ocean_frac.py
@@ -1,5 +1,5 @@
 """pytest for algorithm
-clev2er.algorithms.seaice_stage_2.alg_add_region_mask
+clev2er.algorithms.seaice_stage_3.alg_add_ocean_frac
 """
 
 import logging
@@ -11,7 +11,7 @@ import numpy as np
 import pytest
 from netCDF4 import Dataset  # pylint:disable=no-name-in-module
 
-from clev2er.algorithms.seaice_stage_2.alg_add_region_mask import Algorithm
+from clev2er.algorithms.seaice_stage_3.alg_add_ocean_frac import Algorithm
 from clev2er.utils.config.load_config_settings import load_config_files
 
 logger = logging.getLogger(__name__)
@@ -26,7 +26,7 @@ def config() -> dict:
         dict: config dictionary
     """
     # load config
-    chain_config, _, _, _, _ = load_config_files("seaice_stage_2")
+    chain_config, _, _, _, _ = load_config_files("seaice_stage_3")
 
     # Set to Sequential Processing
     chain_config["chain"]["use_multi_processing"] = False
@@ -78,27 +78,27 @@ merge_file_test = [(0), (1)]
 
 
 @pytest.mark.parametrize("file_num", merge_file_test)
-def test_add_region_mask(
+def test_add_ocean_frac_sar(
     file_num,
     previous_steps: Dict,
     thisalg: Algorithm,  # pylint: disable=redefined-outer-name
 ) -> None:
-    """test alg_add_region_mask.py
+    """test alg_add_ocean_frac.py
 
     Test plan:
     Load a merge file
     run Algorithm.process() on each
     test that the files return (True, "")
-    test that 'region_mask' is in shared_dict, it is an array of bools
+    test that 'ocean_frac' is in shared_dict, it is an array of floats
     """
 
     base_dir = Path(os.environ["CLEV2ER_BASE_DIR"])
     assert base_dir is not None
 
-    # ==================================  FILE TESTING ==========================================
-    logger.info("Testing  file:")
+    # ================================== MERGE FILE TESTING ========================================
+    logger.info("Testing merge file:")
 
-    # load  file
+    # load merge file
     l1b_merge_file = list(
         (base_dir / "testdata" / "cs2" / "l1bfiles" / "arctic" / "merge_modes").glob("*.nc")
     )[file_num]
@@ -114,18 +114,18 @@ def test_add_region_mask(
     for title, step in previous_steps.items():
         success, err_str = step.process(l1b, shared_dict)  # type: ignore[attr-defined]
         if not success:
-            logger.error("Error with previous step: %s\n%s", title, err_str)
+            logger.error(" Error with previous step: %s\n%s", title, err_str)
 
     success, err_str = thisalg.process(l1b, shared_dict)
 
-    assert success, f"Algorithm failed due to: {err_str}"
+    assert success, f" Algorithm failed due to: {err_str}"
 
     # Algorithm tests
-    assert "region_mask" in shared_dict, "'region_mask' not in shared_dict."
+    assert "ocean_frac" in shared_dict, "'ocean_frac' not in shared_dict."
 
     assert isinstance(
-        shared_dict["region_mask"], np.ndarray
-    ), f"'region_mask' is {type(shared_dict['region_mask'])}, not ndarray."
+        shared_dict["ocean_frac"], np.ndarray
+    ), f"'ocean_frac' is {type(shared_dict['ocean_frac'])}, not ndarray."
 
-    mask_dtype = str(shared_dict["region_mask"].dtype)
-    assert "bool" in mask_dtype.lower(), f"Dtype of 'region_mask' is {mask_dtype}, not bool."
+    elev_dtype = str(shared_dict["ocean_frac"].dtype)
+    assert "float" in elev_dtype.lower(), f"Dtype of 'ocean_frac' is {elev_dtype}, not float."
